+++ unittests/structs_minimal.cpp
@@ -24,248 +24,29 @@
   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#include "common.hpp"
-#include <boost/test/unit_test.hpp>
+#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
+#include "structs_minimal.hpp"
 
-class MemberMinimal
-{
-  public:
-    MemberMinimal() = default;
-    MemberMinimal( std::string const & str ) : x( str ) {}
+TEST_SUITE("structs_minimal");
 
-  protected:
-    friend class cereal::access;
-
-    template <class Archive>
-    std::string save_minimal( Archive const & ) const
-    {
-      return x;
-    }
-
-    template <class Archive>
-    void load_minimal( Archive const &, std::string const & str )
-    {
-      x = str;
-    }
-
-  public:
-    std::string x;
-};
-
-class MemberMinimalVersioned
-{
-  public:
-    MemberMinimalVersioned() = default;
-    MemberMinimalVersioned( double d ) : x( d ) {}
-
-  protected:
-    friend class cereal::access;
-
-    template <class Archive>
-    double save_minimal( Archive const &, const std::uint32_t ) const
-    {
-      return x;
-    }
-
-    template <class Archive>
-    void load_minimal( Archive const &, double const & d, const std::uint32_t )
-    {
-      x = d;
-    }
-
-  public:
-    double x;
-};
-
-struct NonMemberMinimal
-{
-  NonMemberMinimal() = default;
-  NonMemberMinimal( std::uint32_t xx ) : x(xx) {}
-  std::uint32_t x;
-};
-
-template <class Archive>
-std::uint32_t save_minimal( Archive const &, NonMemberMinimal const & nmm )
-{
-  return nmm.x;
-}
-
-template <class Archive>
-void load_minimal( Archive const &, NonMemberMinimal & nmm, std::uint32_t const & data )
-{
-  nmm.x = data;
-}
-
-struct NonMemberMinimalVersioned
-{
-  NonMemberMinimalVersioned() = default;
-  NonMemberMinimalVersioned( bool xx ) : x(xx) {}
-  bool x;
-};
-
-template <class Archive>
-bool save_minimal( Archive const &, NonMemberMinimalVersioned const & nmm, std::uint32_t const )
-{
-  return nmm.x;
-}
-
-template <class Archive>
-void load_minimal( Archive const &, NonMemberMinimalVersioned & nmm, bool const & data, std::uint32_t const )
-{
-  nmm.x = data;
-}
-
-struct TestStruct
-{
-  TestStruct() = default;
-  TestStruct( std::string const & s, double d, std::uint32_t u, bool b ) :
-    mm(s), mmv(d), nmm(u), nmmv(b) {}
-
-  template <class Archive>
-  void serialize( Archive & ar )
-  {
-    ar( mm, mmv );
-    ar( nmm, nmmv );
-  }
-
-  MemberMinimal mm;
-  MemberMinimalVersioned mmv;
-  NonMemberMinimal nmm;
-  NonMemberMinimalVersioned nmmv;
-};
-
-struct Issue79Struct
-{
-  Issue79Struct() = default;
-  Issue79Struct( std::int32_t xx ) : x(xx) {}
-  std::int32_t x;
-};
-
-template <class Archive, cereal::traits::DisableIf<std::is_same<Archive, cereal::BinaryOutputArchive>::value ||
-                                                   std::is_same<Archive, cereal::PortableBinaryOutputArchive>::value> = cereal::traits::sfinae>
-std::string save_minimal( Archive const &, Issue79Struct const & val )
-{
-  return std::to_string( val.x );
-}
-
-template <class Archive, cereal::traits::DisableIf<std::is_same<Archive, cereal::BinaryInputArchive>::value ||
-                                                   std::is_same<Archive, cereal::PortableBinaryInputArchive>::value> = cereal::traits::sfinae>
-void load_minimal( Archive const &, Issue79Struct & val, std::string const & str )
-{
-  val.x = std::stoi( str );
-}
-
-template <class Archive, cereal::traits::EnableIf<std::is_same<Archive, cereal::BinaryOutputArchive>::value ||
-                                                  std::is_same<Archive, cereal::PortableBinaryOutputArchive>::value> = cereal::traits::sfinae>
-std::int32_t save_minimal( Archive const &, Issue79Struct const & val )
-{
-  return val.x;
-}
-
-template <class Archive, cereal::traits::EnableIf<std::is_same<Archive, cereal::BinaryInputArchive>::value ||
-                                                  std::is_same<Archive, cereal::PortableBinaryInputArchive>::value> = cereal::traits::sfinae>
-void load_minimal( Archive const &, Issue79Struct & val, std::int32_t const & xx )
-{
-  val.x = xx;
-}
-
-struct Issue79StructInternal
-{
-  Issue79StructInternal() = default;
-  Issue79StructInternal( std::int32_t xx ) : x(xx) {}
-  std::int32_t x;
-
-  template <class Archive, cereal::traits::DisableIf<std::is_same<Archive, cereal::BinaryOutputArchive>::value ||
-                                                     std::is_same<Archive, cereal::PortableBinaryOutputArchive>::value> = cereal::traits::sfinae>
-  std::string save_minimal( Archive const & ) const
-  {
-    return std::to_string( x );
-  }
-
-  template <class Archive, cereal::traits::DisableIf<std::is_same<Archive, cereal::BinaryInputArchive>::value ||
-                                                     std::is_same<Archive, cereal::PortableBinaryInputArchive>::value> = cereal::traits::sfinae>
-  void load_minimal( Archive const &, std::string const & str )
-  {
-    x = std::stoi( str );
-  }
-
-  template <class Archive, cereal::traits::EnableIf<std::is_same<Archive, cereal::BinaryOutputArchive>::value ||
-                                                    std::is_same<Archive, cereal::PortableBinaryOutputArchive>::value> = cereal::traits::sfinae>
-  std::int32_t save_minimal( Archive const & ) const
-  {
-    return x;
-  }
-
-  template <class Archive, cereal::traits::EnableIf<std::is_same<Archive, cereal::BinaryInputArchive>::value ||
-                                                    std::is_same<Archive, cereal::PortableBinaryInputArchive>::value> = cereal::traits::sfinae>
-  void load_minimal( Archive const &, std::int32_t const & xx )
-  {
-    x = xx;
-  }
-};
-
-template <class IArchive, class OArchive>
-void test_structs_minimal()
-{
-  std::random_device rd;
-  std::mt19937 gen(rd());
-
-  for(int ii=0; ii<100; ++ii)
-  {
-    TestStruct o_struct = { random_basic_string<char>(gen), random_value<double>(gen),
-                            random_value<std::uint32_t>(gen), random_value<uint8_t>(gen) % 2 ? true : false };
-
-    Issue79Struct o_struct2 = { random_value<std::int32_t>(gen) };
-    Issue79StructInternal o_struct3 = { random_value<std::int32_t>(gen) };
-
-    std::ostringstream os;
-    {
-      OArchive oar(os);
-      oar( o_struct );
-      oar( o_struct2 );
-      oar( o_struct3 );
-    }
-
-    decltype(o_struct) i_struct;
-    decltype(o_struct2) i_struct2;
-    decltype(o_struct3) i_struct3;
-
-    std::istringstream is(os.str());
-    {
-      IArchive iar(is);
-      iar( i_struct );
-      iar( i_struct2 );
-      iar( i_struct3 );
-    }
-
-    BOOST_CHECK(o_struct.mm.x == i_struct.mm.x);
-    BOOST_CHECK_CLOSE(o_struct.mmv.x, i_struct.mmv.x, 1e-5);
-
-    BOOST_CHECK(o_struct.nmm.x == i_struct.nmm.x);
-    BOOST_CHECK(o_struct.nmmv.x == i_struct.nmmv.x);
-
-    BOOST_CHECK(o_struct2.x == i_struct2.x);
-
-    BOOST_CHECK(o_struct3.x == i_struct3.x);
-  }
-}
-
-BOOST_AUTO_TEST_CASE( binary_structs_minimal )
+TEST_CASE("binary_structs_minimal")
 {
   test_structs_minimal<cereal::BinaryInputArchive, cereal::BinaryOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( portable_binary_structs_minimal )
+TEST_CASE("portable_binary_structs_minimal")
 {
   test_structs_minimal<cereal::PortableBinaryInputArchive, cereal::PortableBinaryOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( xml_structs_minimal )
+TEST_CASE("xml_structs_minimal")
 {
   test_structs_minimal<cereal::XMLInputArchive, cereal::XMLOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( json_structs_minimal )
+TEST_CASE("json_structs_minimal")
 {
   test_structs_minimal<cereal::JSONInputArchive, cereal::JSONOutputArchive>();
 }
+
+TEST_SUITE_END();
